+++ ta4j-core/src/test/java/org/ta4j/core/indicators/averages/VIDYAIndicatorTest.java
@@ -21,51 +21,41 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.mocks;
+package org.ta4j.core.indicators.averages;
 
-import java.time.Duration;
-import java.time.Instant;
+import static org.ta4j.core.TestUtils.*;
 
-import org.ta4j.core.BaseBar;
-import org.ta4j.core.bars.TimeBarBuilder;
+import org.junit.Test;
+import org.ta4j.core.BarSeries;
+import org.ta4j.core.CsvTestUtils;
+import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.AbstractIndicatorTest;
+import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
+import org.ta4j.core.mocks.MockIndicator;
+import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-public class MockBarBuilder extends TimeBarBuilder {
+public class VIDYAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
 
-    private final Instant beginTime = Instant.EPOCH;
-    private boolean periodSet;
-    private boolean endTimeSet;
-
-    private static long countOfProducedBars;
-    private Duration timePeriod;
-
-    public MockBarBuilder(NumFactory numFactory) {
+    public VIDYAIndicatorTest(NumFactory numFactory) {
         super(numFactory);
     }
 
-    @Override
-    public TimeBarBuilder endTime(final Instant endTime) {
-        endTimeSet = true;
-        return super.endTime(endTime);
-    }
+    @Test
+    public void vidyaIndicatorTest() {
 
-    @Override
-    public TimeBarBuilder timePeriod(final Duration timePeriod) {
-        periodSet = true;
-        this.timePeriod = timePeriod;
-        return super.timePeriod(this.timePeriod);
-    }
+        MockIndicator mock = CsvTestUtils.getCsvFile(VIDYAIndicatorTest.class, "VIDYA.csv", numFactory);
 
-    @Override
-    public BaseBar build() {
-        if (!periodSet) {
-            timePeriod(Duration.ofDays(1));
+        BarSeries barSeries = mock.getBarSeries();
+
+        VIDYAIndicator vidya = new VIDYAIndicator(new ClosePriceIndicator(barSeries), 9, 20);
+
+        for (int i = 0; i < barSeries.getBarCount(); i++) {
+            Num expected = mock.getValue(i);
+            Num value = vidya.getValue(i);
+
+            assertNumEquals(expected.doubleValue(), value);
         }
-
-        if (!endTimeSet) {
-            endTime(beginTime.plus(timePeriod.multipliedBy(++countOfProducedBars)));
-        }
-        return super.build();
     }
 
 }